+++ test/AWS.OpenTelemetry.CloudWatch.Plugin.Tests/MetricsExporterEnvironment.cs
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-namespace AWS.OpenTelemetry.CloudWatch.Plugin.Tests;
+namespace AWS.OpenTelemetry.CloudWatchPluginOtel.Tests;
 
 internal sealed class MetricsExporterEnvironment : IDisposable
 {