+++ test/AWS.OpenTelemetry.CloudWatch.Plugin.Tests/SpanMetricsTracerProviderBuilderExtensionsTests.cs
@@ -3,12 +3,12 @@
 
 using System.Diagnostics;
 using System.Reflection;
-using AWS.OpenTelemetry.CloudWatch.Plugin.Tests.Implementation.SpanMetrics;
+using AWS.OpenTelemetry.CloudWatchPluginOtel.Tests.Implementation.SpanMetrics;
 using OpenTelemetry;
 using OpenTelemetry.Metrics;
 using OpenTelemetry.Trace;
 
-namespace AWS.OpenTelemetry.CloudWatch.Plugin.Tests;
+namespace AWS.OpenTelemetry.CloudWatchPluginOtel.Tests;
 
 [Collection(SpanMetricsTestsCollection.Name)]
 public class SpanMetricsTracerProviderBuilderExtensionsTests