+++ src/AWS.OpenTelemetry.CloudWatch.Plugin/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+* Moved `service.name` from metric datapoint attributes to the metric resource
+  and set the `traces.span.metrics.calls` unit to `{call}`.
+* Aligned the assembly and namespaces with the
+  `AWS.OpenTelemetry.CloudWatchPluginOtel` package name.
+
 ## 0.1.0 - 2026-08-24
 
 * Added span metrics.